Could tell if --bench in argv but fail in inf loop
Call both the programs with argument --bench s "key word" enable
the program to search "key word" int prefix.
The current program can search the right cities that user assigned,
but it will enter a infunute loop which keep showing the
result of search.
The problem might be in the for loop of tst.c

Modify the "test" in Makefile. "test" can call teh perf to
test the efficiency of the program. But htere is sth went wrong.

Modify the "astyle" in  Makefile. Make the fomation more similar
to the git accepted one.