Add memory pool successfully
Add the memory pool to the testMpool. And it seems that
the program run successfully. But I have to modify the size
of the memory pool when Im using "mPool_allocate" func.
the size should be well calculated.

There is a bug in Makefile. the cache-test_mpool has
wrong directory, so the cache-test cant run successfully.
Need to fix the bug.

Modidy the .gitignore which let git to ignore Makefile.swo
and Makefile.swp.